eval/pyproject.toml
[project]
name = "shopchat-eval"
version = "0.1.0"
description = "Golden set evals for ShopChat AI assistant"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pyyaml>=6.0.0",
    "langfuse>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["."]
